update config file for nw
