+++ example/chat-app/.gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 node_modules/
 /dist/
+/dev_dist/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
