+++ example/chat-app/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <title>chat-app</title>
+    <script type="text/javascript" src="./Carrier/index.js"></script>
   </head>
   <body>
     <div id="app"></div>
