+++ example/chat-app/src/utility/Log.js
